+++ src/queue_server.cpp
@@ -11,8 +11,17 @@
 #include "value_type.h"
 #include "server.h"
 
+#include "statistics.h"
+
 class queue_server {
+public:
+    enum {
+        into_any = 1,
+        into_hash = 2
+    };
+private:
     struct instruction {
+
         instruction() = default;
         instruction(const instruction&) = default;
         instruction(instruction&&) = default;
@@ -21,6 +30,7 @@ class queue_server {
         heap::small_vector<uint8_t, 32> value{};
         art::tree*  t{};
         art::key_options options{};
+        int into = into_any;
         void set_key(art::value_type k) {
             key.append(k.to_view());
         }
@@ -33,7 +43,31 @@ class queue_server {
         [[nodiscard]] art::value_type get_value() const {
             return {value.data(), value.size()};
         }
+        [[nodiscard]] bool exec() const {
+            try {
+                ++statistics::queue_processed;
+                if (!t) {
+                    ++statistics::queue_failures;
+                    return false;
+                }
+                write_lock release(t->latch);
+                --t->queue_size;
+                // TODO: the infernal thread_ap should be set before using any t functions
+                if (into == into_hash) {
+                    t->hash_insert(options, get_key(),get_value(),true,[](const art::node_ptr& ){});
+                    return true;
+                }
+                if (into == into_any) {
+                    t->opt_insert(options, get_key(),get_value(),true,[](const art::node_ptr& ){});
+                    return true;
 
+                }
+            }catch (std::exception& e) {
+                art::std_err("exception processing queue", e.what());
+            }
+            ++statistics::queue_failures;
+            return false;
+        }
     };
     public:
     queue_server() {
@@ -56,14 +90,7 @@ class queue_server {
             while (started) {
                 instruction ins;
                 if (q->try_dequeue(ins)) {
-                    try {
-                        write_lock release(ins.t->latch);
-                        --ins.t->queue_size;
-                    // TODO: the infernal thread_ap should be set before using any t functions
-                        ins.t->opt_insert(ins.options, ins.get_key(),ins.get_value(),true,[](const art::node_ptr& ){});
-                    }catch (std::exception& e) {
-                        art::std_err("exception processing queue", e.what());
-                    }
+                    ins.exec();
                 }else {
                     std::this_thread::sleep_for(std::chrono::milliseconds(10));
                 }
@@ -78,23 +105,24 @@ class queue_server {
 
 
     }
-    void queue_insert(art::tree* t,art::key_options options,art::value_type k, art::value_type v) {
+    void queue_insert(art::tree* t,art::key_options options,art::value_type k, art::value_type v, int where) {
+
         size_t at = t->shard % threads.size();
         instruction ins;
+        ins.into = where;
         ins.set_key(k);
         ins.set_value(v);
         ins.options = options;
         ins.t = t;
         ++t->queue_size;
         queues[at]->enqueue(ins);
+        ++statistics::queue_added;
     }
     void consume() {
         for (auto& q : queues) {
             instruction ins;
             while (q->try_dequeue(ins)) {
-                write_lock release(ins.t->latch);
-                --ins.t->queue_size;
-                ins.t->opt_insert(ins.options, ins.get_key(),ins.get_value(),true,[](const art::node_ptr& ){});
+                ins.exec();
             }
         }
     }
@@ -107,9 +135,7 @@ class queue_server {
         // of the actual queue size - but usually it's the same
         while (t->queue_size > 0) {
             if (queues[q]->try_dequeue(ins)) {
-                write_lock r1(ins.t->latch);
-                --ins.t->queue_size;
-                ins.t->opt_insert(ins.options, ins.get_key(),ins.get_value(),true,[](const art::node_ptr& ){});
+                ins.exec();
             }else {
                 std::this_thread::yield();
             }
@@ -144,8 +170,18 @@ bool is_queue_server_running() {
     return server != nullptr;
 }
 
+void hash_queue_insert(size_t shard, art::key_options options,art::value_type k, art::value_type v) {
+    auto t = get_art(shard);
+    if (server)
+        server->queue_insert(t,options,k,v,queue_server::into_hash);
+    else
+        t->hash_insert(options,k,v,true,[](const art::node_ptr& ){});
+}
+
 void queue_insert(size_t shard, art::key_options options,art::value_type k, art::value_type v) {
     auto t = get_art(shard);
     if (server)
-        server->queue_insert(t,options,k,v);
+        server->queue_insert(t,options,k,v,queue_server::into_any);
+    else
+        t->opt_insert(options,k,v,true,[](const art::node_ptr& ){});
 }